api/pyproject.toml
[project]
name = "lumina-api"
version = "0.1.0"
description = "FastAPI backend + HTMX UI for LED strip controller"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "paho-mqtt>=2.1.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "bcrypt>=4.1.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.0",
    "itsdangerous>=2.1.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]
